For TimeZoneDatabase, remove some lists Instead of returning lists with two periods or two DateTime structs, take them out of the list and put them directly in the tuple.
